Fix tear-down error and optimize clean up

+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/tests/SecretsTestBase.cs
@@ -11,6 +11,7 @@ using NUnit.Framework;
 
 namespace Azure.Security.KeyVault.Secrets.Tests
 {
+    [NonParallelizable]
     public abstract class SecretsTestBase : RecordedTestBase
     {
         public const string AzureKeyVaultUrlEnvironmentVariable = "AZURE_KEYVAULT_URL";
@@ -21,7 +22,9 @@ namespace Azure.Security.KeyVault.Secrets.Tests
 
         public Uri VaultUri { get; set; }
 
-        private readonly ConcurrentQueue<string> _secretsToCleanup = new ConcurrentQueue<string>();
+        // Queue deletes, but poll on the top of the purge stack to increase likelihood of others being purged by then.
+        private readonly ConcurrentQueue<string> _secretsToDelete = new ConcurrentQueue<string>();
+        private readonly ConcurrentStack<string> _secretsToPurge = new ConcurrentStack<string>();
 
         protected SecretsTestBase(bool isAsync) : base(isAsync)
         {
@@ -46,30 +49,51 @@ namespace Azure.Security.KeyVault.Secrets.Tests
             VaultUri = new Uri(Recording.GetVariableFromEnvironment(AzureKeyVaultUrlEnvironmentVariable));
         }
 
-        [OneTimeTearDown]
+        [TearDown]
         public async Task Cleanup()
         {
-            List<Task> cleanupTasks = new List<Task>();
-
-            foreach (string name in _secretsToCleanup)
+            // Start deleting resources as soon as possible.
+            while (_secretsToDelete.TryDequeue(out string name))
             {
-                Task cleanupTask = CleanupSecret(name);
-                cleanupTasks.Add(cleanupTask);
-            }
+                await DeleteSecret(name);
 
-            await Task.WhenAll(cleanupTasks);
+                _secretsToPurge.Push(name);
+            }
         }
 
-        protected async Task CleanupSecret(string name)
+        [OneTimeTearDown]
+        public async Task CleanupAll()
         {
+            // Make sure the delete queue is empty.
+            await Cleanup();
+
+            while (_secretsToPurge.TryPop(out string name))
+            {
+                await PurgeSecret(name).ConfigureAwait(false);
+            }
+        }
+
+        protected async Task DeleteSecret(string name)
+        {
+            if (Mode == RecordedTestMode.Playback)
+            {
+                return;
+            }
+
             try
             {
-                await Client.StartDeleteSecretAsync(name).ConfigureAwait(false);
+                using (Recording.DisableRecording())
+                {
+                    await Client.StartDeleteSecretAsync(name).ConfigureAwait(false);
+                }
             }
             catch (RequestFailedException ex) when (ex.Status == 404)
             {
             }
+        }
 
+        protected async Task PurgeSecret(string name)
+        {
             try
             {
                 await WaitForDeletedSecret(name).ConfigureAwait(false);
@@ -78,9 +102,17 @@ namespace Azure.Security.KeyVault.Secrets.Tests
             {
             }
 
+            if (Mode == RecordedTestMode.Playback)
+            {
+                return;
+            }
+
             try
             {
-                await Client.PurgeDeletedSecretAsync(name).ConfigureAwait(false);
+                using (Recording.DisableRecording())
+                {
+                    await Client.PurgeDeletedSecretAsync(name).ConfigureAwait(false);
+                }
             }
             catch (RequestFailedException ex) when (ex.Status == 404)
             {
@@ -89,7 +121,7 @@ namespace Azure.Security.KeyVault.Secrets.Tests
 
         protected void RegisterForCleanup(string name)
         {
-            _secretsToCleanup.Enqueue(name);
+            _secretsToDelete.Enqueue(name);
         }
 
         protected void AssertSecretsEqual(KeyVaultSecret exp, KeyVaultSecret act)
